+++ CHANGELOG.md
@@ -2,4 +2,14 @@
 
 ## Unreleased
 
-- TODO
+- Add a [<q>run</q>](https://code.visualstudio.com/api/extension-guides/virtual-documents) button to directly run the Game.exe.
+- Automatically close all tabs when running `Close Scripts.rvdata2` command.
+
+## 0.1.1
+
+- Add a [tree view](https://code.visualstudio.com/api/extension-guides/tree-view) as a replacement for the file explorer.
+
+## 0.1.0
+
+- Add a [file system provider](https://github.com/microsoft/vscode-extension-samples/tree/main/fsprovider-sample) which can open and edit entries in Scripts.rvdata2.\
+  Many thanks to the [ZipFS](https://github.com/yarnpkg/berry/tree/master/packages/vscode-zipfs) extension written by Yarn.